Vastly simplify the geometry of Noto flags. Replace the complex path for the wave with a single cubic Bezier curve.
This should make it easier to analytically distort more complex images than stripes, paving the way to non-stripe flags.
The only remaining use of the complex path is the inside of the flag's border, which I may replace after working through accurately truncating the curve.

Unlike the rainbow flag emoji, the region flags in Noto Emoji seem to be raster flag images acquired from Wikipedia, processed through a waveflag.c script which introduces the wave to the flag. There definitely seems to be some post-processing or modification of the script as configured by default; the final images (as viewed on, e.g. the Noto Emoji website) have a 4 pixel margin on each x side instead of 1 pixel, have aligned beginnings and ends of the wave, etc. Instead of working through the accurate transformation process to calculate an exact curve, I found an empirical match. Assuming purely linear postprocessing, the wave would still be cubic with control points mirrored across a center point, and assuming there are no rotations or horizontal skews (which would distort the vertical edges of the flag), those control points are exactly 1/3 and 2/3 of the way horizontally across the wave. This leaves only one degree of freedom, the y offset of the control points, for which 24.1 produces incredibly close results. There may be a better value, but any further adjustments would likely be overwhelmed by the rounding errors in the original, complex path.